Move the location of the mysql db (used by phabricator) to /home/ubuntu
Summary:
/home/ubuntu is a big disk, that has lots of space, unlike the root
disk.  I think in ec2-land it's probably not a good idea to put any
user data on the disk holding /.  This is one step in that direction.

I had to make two changes: one to the mysql config and another to
apparmor to let it know about the new location.  Props to the mysql
config file for commenting the need to change apparmor!

Reviewers: chris, benkomalo

Differential Revision: http://phabricator.khanacademy.org/D69