Convert analytics setup instructions to a script
Summary:
There are still lots of TODO's here, including setting up postfix
properly and testing the crontab. I ran the script and it seems to do
everything else properly (though I did not actually install the crontab)

I also have not actually tried running the jobs quite yet. I'm not
convinced those PYTHONPATH's are correct (or even needed...I'd like for
that not to be required)

Reviewers: csilvers

Reviewed By: csilvers

CC: jace, yunfangjuan

Differential Revision: http://phabricator.khanacademy.org/D83