Rework the README into a shell script.
Summary:
In theory, you should be able to just run the shell script to set up a
new machine, but it's not nearly that easy, since most of the services
here require steps that are difficult to automate (e.g., setting up a
jenkins admin account, which requires going to a webpage on the
jenkins server).  I've punted on most of these.  Probably the right
way to resolve these is via puppet or an AMI image or something.

However, it has all the information in the README, as well as in
gerrit.install, so it's an improvement at least...

As the script says:

Reviewers: benkomalo, chris

Reviewed By: benkomalo

Differential Revision: http://phabricator.khanacademy.org/D96