Fix mongo_ctrl so it can be used in /etc/init.d
Summary:
I broke this while cleaning it up the other day; it still works when
executed normally, but we should consider making it a real service
(which it currently isn't running as).

A big TODO is to figure out who should own the processes/db files; right
now the 'analytics' user owns the mongo files and processes. I think if
we install this into init.d, it'll run as root, which may affect
ownership of the files.

Reviewers: csilvers, yunfangjuan

Reviewed By: yunfangjuan

Differential Revision: http://phabricator.khanacademy.org/D107