Use bash --login to load the .bash_profile.  That's the easiest to debug.
Summary:
We were having trouble with the cronjobs running via sh (which I think
is ash, or maybe dash), rather than bash.  When we tried to source the
.bash_profile, it would give an error because of all the bash-isms in
it.

The simplest solution is to just run everything under bash.  If we use
the --login flag, it will load .bash_profile, just like it does when
one logs into the analytics machine.  That makes it so the cronjob is
running in as close to the same environment as a manual run, as
possible.

Hopefully this should fix the crontab mail once and for all.  Latest
was:
---
From: root@khanacademy.org (Cron Daemon)
To: toby-admin+ka-high-mem@khanacademy.org
Subject: [toby-admin] Cron <analytics@ip-10-0-0-108> . ~/.bash_profile;
 ~/analytics/src/gae_download.py -c ~/analytics/cfg/gae_download.json  (failed)

/home/analytics/.bashrc: 13: shopt: not found
/home/analytics/.bashrc: 21: shopt: not found
/home/analytics/.bashrc: 103: shopt: not found
/etc/bash_completion: 32: [[: not found
/etc/bash_completion: 38: [[: not found
/etc/bash_completion: 51: Bad substitution
---

While in the area, I also fixed up a typo I noticed: % has to be
escaped in crontabs.

Be explicit about what shell we're using.  Requested by jace.

Reviewers: yunfangjuan, jace

Reviewed By: jace

Differential Revision: http://phabricator.khanacademy.org/D124