Add script to create a role account
Summary: Let me know if the way I have this set up with echoing stuff makes sense -- I was trying to avoid having to copy the script while at the same time having the username be a command line argument. Maybe it makes more sense to have the script prompt for the username.

Test Plan: Ran it on a new instance.

Reviewers: benkomalo

Reviewed By: benkomalo

Differential Revision: http://phabricator.khanacademy.org/D344