Add in skeletal setup files for reports machine
Summary:
This will be the machine that replaces ka-backup. I hope to set up
the reporting db (that was originally on ka-backup, but can no longer
run since ka-backup is 32-bit...) here, and then eventually the actual
dashboard code itself so it can read off the local mongo instance and
we'd be able to close off those default mongo ports.

Reviewers: jace, david

Reviewed By: jace

CC: yunfangjuan

Differential Revision: http://phabricator.khanacademy.org/D343