Fix the install by making this an absolute symlink, not relative.
Summary:
The install does a 'cp -s' of this symlink, which means we ended up with
   /etc/lighttpd/conf-enabled/15-fastcgi-php.conf ->
   aws-config/internal-webserver/etc/lighttpd/conf-enabled/15-fastcgi-php.conf ->
   ../conf-available/15-fastcgi-php.conf

So this ends up resolving to
   aws-config/internal-webserver/etc/lighttpd/conf-available/15-fastcgi-php.conf

But this file doesn't exist!  The conf-available/ dir is a system
directory, controlled by apt-get and the like.  By changing the
symlink to absolute, we get
   /etc/lighttpd/conf-enabled/15-fastcgi-php.conf ->
   aws-config/internal-webserver/etc/lighttpd/conf-enabled/15-fastcgi-php.conf ->
   /etc/lighttpd/conf-available/15-fastcgi-php.conf
which works fine.

Reviewers: benkomalo

Reviewed By: benkomalo

Differential Revision: http://phabricator.khanacademy.org/D544