Fetch GAE usage reports on analytics machine
Summary:
This cron job runs daily and invokes a script that lives in the
analytics git repo.

Test Plan:
I installed the crontab entry on analytics and set it for a time to run,
then verify that the output was as expected and the logs were written to
the right place.

Reviewers: benkomalo

Reviewed By: benkomalo

Differential Revision: http://phabricator.khanacademy.org/D861