Raise the timeout on fetch_logs.sh
Summary:
It's taking more than an hour to fetch an hour's worth of logs,
sometimes.  It's not a problem to have the log-fetches overlap, so we
up the timeout.

We want to keep *some* timeout so hanging jobs don't stick around for
forever.

Make the timeout 120 min, not 180.

The load_emr.sh script runs at 2:30 AM UTC to access the prior day's
data. With a 3h timeout, it's possible that the final run (11pm-12am
UTC) might not be complete by the time that the log processing occurs,
which might make for a difficult to track down error.

Reviewers: chris

Reviewed By: chris

Differential Revision: http://phabricator.khanacademy.org/D879