Update init scripts on the ci box for the 12.04 upgrade.
Summary:
While diagnosing an issue with exercise-screens-daemon, I decided to take the
opportunity to upgrade the ci box to Ubuntu 12.04.1, since the version of
Upstart that ships with it has better logging.

Exercise-screens was creating zombie processes because it couldn't find
phantomjs in its path, so I now explictly specify the path in the job
description. (Likewise updated the job for Mr. so it can find the GAE SDK and
$HOME.)

Test Plan: on the ci box itself

Reviewers: dylan, csilvers

Reviewed By: csilvers

CC: alpert, stephanie

Differential Revision: http://phabricator.khanacademy.org/D2082