add nobootwait to entries in fstab
Summary:
This part of the script doesn't run, so this is essentially a
documentation change.  But it seems a good idea to include
the nobootwait flag, otherwise if there is difficulty mounting
a drive on instance reboot, the machine will hang and wait for
interactive input on what to do.  noboootwait will allow the
instance to come up, so you can then ssh in and fix things up.

Test Plan:
I verified this gets you past bad mounts, because
I had to start to detach the root device for the analytics
machine, start up new EC2 instance and attach the old device,
which let me edit the fstab, then reattached and rebooted
the analytics machine and it came up.  Whew.

Reviewers: chris, csilvers

Reviewed By: csilvers

CC: csilvers, benkomalo

Differential Revision: http://phabricator.khanacademy.org/D2117