add nobootwait to fstab.extra
Summary:
per your request.  I checked and 'df -h' on toby showed
and additional 400GB drive mounted.  I suppose that is the free
instance storage.  I don't know if that should also be added
to the fstab, though.  I think is supposed to survive a reboot,
but not a termination.

Test Plan: none-- similar changes were tested for D2117

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: http://phabricator.khanacademy.org/D2118