Start sending snapshot mail only on error, and increase # of snapshots.
Summary:
I've been monitoring the system every day for the last 6 weeks or so,
and it's been working perfectly.  The current snapshot story on aws is
this:
   2013-03-01 04:00 PST
   2013-03-17 05:00 PDT
   2013-03-24 05:00 PDT
   2013-03-31 05:00 PDT
   2013-04-01 05:00 PDT
   2013-04-03 05:00 PDT
   2013-04-04 05:00 PDT
   2013-04-07 05:00 PDT
   2013-04-08 05:00 PDT
   2013-04-09 05:00 PDT
   2013-04-10 05:00 PDT

whic is just what we want: 2 monthly snapshots, 4 weekly snapshots,
and 6 daily snapshots.  (This doesn't add up to 16 because we are
leaving space for 3 more monthly snapshots, and we have some slack
because of integer division in dividing up snapshots.)

I'm upping the number of snapshots we keep by a factor of 4, now that
everything is tested to be working IRL.  Snapshots are cheap.

Reviewers: benkomalo

Reviewed By: benkomalo

Differential Revision: http://phabricator.khanacademy.org/D2134