Set up an nginx proxy to Jenkins
Summary:
According to the internet, this sounds like the right way to get access
to Jenkins on port 80.  Using this configuration, recommended by the
Jenkins site, means that static files will be smartly handled by nginx.
This isn't critical, but seems nice, and was easy to copy/paste.

Test Plan: Ran these changes on the jenkins server, and they worked!

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: http://phabricator.khanacademy.org/D2166