Fix permissions issue creating secrets directory
Summary:
Oops, I needed to be the jenkins user to create this directory and write
to it, and I can't do this until after jenkins is installed. I also
modified the printed message at the end of the script to be clear about
the permissions when editing the secrets password file.

Test Plan: Ran aws-config/jenkins/setup.sh on jenkins.khanacademy.org

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: http://phabricator.khanacademy.org/D2535