Use a custom version of the hipchat plugin
Summary:
This custom version has two benefits:
  1) passwords are stored outside of the configuration XML.
  2) when the ping to HipChat fails, the HTTP response body will be
     shown in the logs.

I added instructions that describe how to get this custom plugin
version, and removed the bit from the setup script that would overwrite
it with an older, less useful version of the plugin.

Ignore the fact that the new plugin's version is 0.1.2-SNAPSHOT and the
old version is 0.1.4. This is due to some oddness is how the plugins are
built for Jenkins. The new plugin should really be 0.1.5-SNAPSHOT.

Test Plan:
I ran the steps added to setup.sh on my local development machine,
copied hipchat.hpi to jenkins.khanacademy.org and used some temporary
configuration to turn on HipChat notifications for a test build.

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: http://phabricator.khanacademy.org/D2596