New jobs: deploy-staging and deploy-default
Summary:
These jobs run deployments to production. deploy-staging is run
automatically by a build trigger when the website-commit job succeeds.

This change also introduces the use of a reference repository for the
webapp project. With it git can pull commits from the local disk rather
than over the network. It really speeds up the first build on an empty
workspace, which is common for a new job or when debugging by deleting
the workspace.

Test Plan:
This config is currently running live on jenkins.khanacademy.org, I just
ran the fetch script to sync it into source control:

 $ ./fetch_jenkins_config.sh

Reviewers: csilvers, desmond

Reviewed By: desmond

Differential Revision: http://phabricator.khanacademy.org/D2727