Rename job deploy-default -> deploy-daily
Summary:
This also adds HipChat notification when the build is started, but not
when it ends because tools/build-deploy.sh handles that.

Test Plan:
I ran the job on jenkins.khanacademy.org far enough to begin deploy.py
without errors. I had to rename the absolute paths in core.worktree for
each submodule due to using git 1.7.9 (see http://stackoverflow.com/questions/10666847/how-to-rename-a-git-repo-project-which-contains-submodules-in-its-subdirectori):

  $ cd webapp
  $ find .git/modules/ -name config -print0 | xargs -0 -n1 sed
--in-place= s/deploy-default/deploy-daily/

I also had to delete the virtualenv directory and allow it to rebuild
because many .so and python scripts referenced the previous path:

  $ rm -rf env
  # Then I kicked off a build from the web UI but killed it after it
  # rebuilt the virtualenv.

Reviewers: desmond

Reviewed By: desmond

Differential Revision: http://phabricator.khanacademy.org/D2729