Sync config with jenkins.khanacademy.org
Summary:
Switch from running "git clean ..." in Jenkins to making it part of
build-deploy.sh as the CLEAN variable.  We do now expose this value as a
parameter to the build.  Usually, the default of "most" is what you
want, for a fast build.  Sometimes, you want "all" to ensure everything
is cleared out.

website-commit no longer polls for changes every minute, but expects
Kiln to call the git-plugin's notifyCommit webhook.  I had to set
website-allcheck to ignore the notifyCommit call, since otherwise it
would also be triggered by the webhook.

I added post-build tasks to website-commit that look for records of lint
errors or test failures and notify HipChat with them.  The idea is to
quickly have an idea of what is failing.  It would be even better if the
test failures contained links. Maybe later.

Test Plan: This code is already running live.

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: http://phabricator.khanacademy.org/D2801