Bring Jenkins config up-to-date
This improves the daily deply workflow. The idea now is to allow a
commit to successfully build website-commit and deploy-staging, and then
to promote it manually from deploy-staging to deploy-daily. This way,
you have a known-good commit that is ok to FORCE_INDEX_CHANGES and
SKIP_TESTS. So, the promotion XML file is new, and deploy-staging is
modified, and deploy-daily now support GIT_COMMIT_OVERRIDE.

This also fixes bad CSS in the Promotions plugin. Previously, long
parameter descriptions would be "pre" formatted and not wrap at the edge
of the screen. "white-space:normal" fixes this.

Dylan is now an admin (welcome back!).

Calls to hipchat_room_message now escape incoming HTML so as not to
confuse HipChat.

Auditors: desmond

Test Plan:
This is running live on jenkins.khanacademy.org