More Jenkins config changes from plugin upgrades
It's OK to gloss over most of these XML changes, I added an auditor so
that someone is aware of what a Jenkins plugin upgrade looks like in
terms of config changes (lots of confusing and mostly harmless XML updates).

After upgrading plugins, it looks like I also needed to re-save the
global config to have it auto-update its XML format.

The obnoxious part is the new format of the disk usage plugin, which
appears to store its data in the job's config XML, and that data is
updated pretty frequently.

Auditors: csilvers

Test Plan:
Already running live on jenkins.khanacademy.org