extend the timeout on the hourly fetch_logs process
We rarely but occasionally get timeouts in the process, which
often leads to a corrupt gzip file as the process gets killed
while gzipping, and that creates failures down the line.
When analytics-admin email recives something like this, then
we've had such a failure:

to analytics-admi.
analytics: fetch-logs failed with error 124 [redo-it manually?]

We've had two such failures in the last couple weeks.  This
change'll give us a little more breathing room.

Test Plan:
fingers crossed

Auditors: chris