Modify setup.sh to use /mnt the way we actually do on the jenkins machine.
Summary:
I also ran

   diff -ur --exclude jobs jenkins/jenkins_home/ /var/lib/jenkins/

and fixed up all the differences I saw, so that this repo now matches
what's actually on jenkins.

(There were diffs in the jobs directory too, but I didn't understand
them as well, so left them alone.  Mostly they're of the form:

-      <diskUsageWithoutBuilds>3203019797</diskUsageWithoutBuilds>
+      <diskUsageWithoutBuilds>9457073014</diskUsageWithoutBuilds>

and I don't really understand what a stat like that is doing in a
config file, so I'm not going to mess with it.)

Test Plan: Fingers crossed.

Reviewers: chris

Reviewed By: chris

Differential Revision: http://phabricator.khanacademy.org/D7963