Adds Amazon CloudSearch configuration script.
Summary:
I created a script that will read a YAML configuration file and
provision the domain from it. I believe this will be more maintainable
into the future, rather than just writing a massive shell script to
run all of the commands (which I initially did but stopped after it
got too difficult to meaningfully deal with).

Test Plan:
Repeatedly provisioned the khan-academy-dev domain with the included
configuration.

Reviewers: benkomalo

Reviewed By: benkomalo

Differential Revision: http://phabricator.khanacademy.org/D8125