Added support for stopwords and alg stemming.
Summary:
You can now specify a stopwords file in the configuration that
contains a YAML list of stopwords to assign to an analysis scheme. You
can also specify a level of algorithmic stemming.

This change also adds two new flags to the provisioning tool that aid
in testing. `--dry-run` which prevents any changes from occuring, but
still prints out what it would do. And `--leave-temp-dir` that disables
automatic deletion of a temporary directory that is now created (the
temp directory was added because of the stopwords change).

Test Plan:
- Ran provisioning script on new domain and observed that everything
  was set up properly.
- Verified that the set of queries {"how is the world", "how is world",
  "how world", "world"} all return identical results. "wow", "is", and
  "the" are all stopwords.
- Verified that "emergency" and "emergencies" both return identical
  results.

Reviewers: benkomalo

Reviewed By: benkomalo

Maniphest Tasks: T2419

Differential Revision: http://phabricator.khanacademy.org/D8294