Allow reading other conf.d files I may write.
I had written a new one to fix
   http://phabricator.khanacademy.org/config/issue/mysql.mode/
but it wasn't taking effect, so strange and hard to undertand, and
wouldn't you know it apparmor was the culprit.

I figured this out when adding the config option to ubuntu_rootdir.cnf
worked but putting it in a new file didn't.  A grep turned up this
culprit.

Auditors: alpert

Test Plan:
Fingers crossed.