Increase fetch_logs.sh timeout, 5 hours -> 7 hours
Not great. @jace recently increased this from 3 hours to 5 hours, but
we're still seeing timeouts (error code 124 is from timeout when the
command it's running times out). More time is required because we put
more in logs now, between logging info about stats and bingo, because
our request volume has increased, and because we recently are
downloading logs from not only the default module but the non-default
modules as well (mapreduce has lots of logs).

Auditors: jace

CC: csilvers

Test Plan:
Fingers crossed.