Fixes problem with aws-proxy CGI script.
Summary:
This was a particularly difficult bug to work out. The problem occurs
because curl likes to use the `Expect: 100-continue` header and wait
for a response before sending its data over. Because we're shoving all
of curl's standard output back to the client (our app on GAE), the
client gets an extra HTTP header, which confuses it greatly.

Since the problem was only reproducable (seemingly) when sending data
through our prod-rpc server, I monkeyed our cgi-script already to
include this patch, and I know it works because I've already tested
it with a publish through production.

Note that this script is only executed during publish time when
publishing to CloudSearch and affects nothing else so I felt very safe
editing the script directly.

A tcpdump I took shows the problem well (the error from CloudSearch is
unrelated, just look at the extra response in the middle):
https://gist.github.com/brownhead/fdeb36389ef41da84684

This change also strips the `ACCEPT-ENCODING: gzip` that GAE likes to
add to its requests because it was annoying when debugging.

This change also adds the verbose flag to curl which gives us header
imformation in the logs (super handy).

Test Plan:
Did a publish in production and observed it go through cleanly by
looking at the logs and confirming that the search results had been
updated.

Reviewers: benkomalo, csilvers, dylan

Reviewed By: csilvers, dylan

Differential Revision: http://phabricator.khanacademy.org/D8661