Use nginx instead of lighttpd
Summary: epriestley suggested that this might improve performance, so let's do it.

Test Plan: I configured the machine manually and it works. Didn't run the script because we're scared to. Went to Phabricator and looked at profiles of pages using DarkConsole; didn't see unnecessary setup checks like I was seeing previously with lighttpd.

Reviewers: johnsullivan, csilvers

Reviewed By: johnsullivan, csilvers

Subscribers: johnsullivan

Differential Revision: http://phabricator.khanacademy.org/D9087