Set max size allowed for POSTs to 8MB
We recently switched from lighttpd to nginx in front of phabricator.
Lighttpd defaults this value to 2GB, which was more than enough. Nginx
defaults to 1MB.

Auditors: alpert
Subscribers: csilvers

Test Plan:
Fingers crossed. I had put this directive in the http, server, and
location contexts and saw that it fixed my issue (uploading 2MB of POST
data), but there's no sense in putting it all places, so I'm guessing
that the server context is the right single place.