Switches production-rpc to nginx.
Summary:
This change includes configuration files for nginx as well as
modifications to the vagrant provisioning code to make the Vagrant box
use nginx as well.

I haven't updated production-rpc yet with the new configuration but I
have tested with the vagrant box. I will update production-rpc after
this review is accepted but before landing.

We lose a nicety that we had with lighttpd which was the configuration
file containing just the production secret. Unfortunately nginx doesn't
have good enough support for variables for us to do this and we have to
edit the nginx configuration file directly instead (or use a templating
engine) to add the secret in. I made mention of this fact at the top of
`install_solr.bash`.

We gain a nicety though... No more CGI script :D!

Test Plan:
Fired up vagrant box and did a full publish with CloudSearch and Solr
enabled. Both reverse proxying configurations (one to the remote
CloudSearch server and one to the local tomcat instance) worked fully.

To ensure that the remote CloudSearch configuration is set up correctly
I hit the vagrant box with curl while running tcpdump and observed
nginx correctly sending out DNS requests in tune with our CloudSearch
domain's TTL.

Reviewers: alpert, csilvers

Subscribers: benkomalo

Differential Revision: http://phabricator.khanacademy.org/D9232