Makes kasandbox.org proxy through to a user-specified host.
Summary:
This will allow us to use kasandbox.org with non-default modules and versions,
which is useful for testing kasandbox.org as well as for serving i18n requests
from the correct module.  Resolves T2661.

This goes with a forthcoming change (D10070) to the client-side code.  Either should be
safe to deploy without the other.

Test Plan:
`varnishd -C -f varnish_default_vcl`, which the internet tells me will check
that Varnish can at least compile the VCL code.
If we deploy this before the client-side code, unless it's massively wrong, it
should work fine for users using the old client code, and allow testing the new
parts of the Varnish config manually.

Reviewers: alpert

Reviewed By: alpert

Subscribers: chris, alan, pamela

Maniphest Tasks: T2661

Differential Revision: http://phabricator.khanacademy.org/D10069