Update .ssh/config to use ReadWriteKiln for all github access.
Before we would only do so when using a weird hostname
(github.com-jenkins).  I'm not sure what the reason for that is -- the
jenkins user only has access to the one ssh key, so might as well just
use it always.  Now that we need that key for subrepos, it's annoying
to use a non-standard hostname for ssh-ing.

Auditors: chris

Test Plan:
On jenkins, went to
   /mnt/jenkins/jobs/update-translations/workspace/webapp/intl/translations
and did
   git checkout master
   git pull
and did not get an 'access denied' error.