Add in a kind ordinal to Cloudsearch domain
Summary:
We wish to tweak with ranking algorithms that rank certain kinds higher
or lower. Unfortunately, the custom search expressions cloudsearch
provides don't allow string literals, making this rather difficult.

This introduces a kind_ordinal, which is only used for ranking purposes.
If it turns out to be successful, it's possible that we could wholly
replace `kind` with `kind_ordinal`, but that is a non goal right now.

Test Plan:
 ran `python provision-domain.py khan-academy-dev-2 domain-info.yaml`
 checked AWS web console and saw the field was created without issues

Reviewers: dylan, alpert

Reviewed By: dylan, alpert

Subscribers: johnsullivan

Differential Revision: http://phabricator.khanacademy.org/D10857