Add memory overcommitting on jenkins.
Deploys and tests regularly run out of memory because they do
fork+exec (to run nodejs, etc), and the process they are forking uses
a lot of memory, so the fork uses the same amount of memory even
though it's immediately replaced by some exec-ed process that uses
almost no memory.  This means we have to have twice as much memory
available as we actually use: suckasaurus.  The solution is to allow
memory overcommitting, which works because the memory in the forked
process is never touched (it's immediately freed when we exec).  cf.
   http://stackoverflow.com/questions/15608347/fork-failing-with-out-of-memory-error

This technique has been working successfully on the make-check-workers
for a few weeks now.

Auditors: chris

Test Plan:
Fingers way crossed.