A bunch of cleanups and improvements to the toby config.
Summary:
1. Change beep-boop and gae-default-version-notifier to use alertlib
2. Run email_bq_data in a daily cronjob
3. Remove unused git_mirrors/hg_mirrors from setup.sh
4. Remove unused install_gerrit from setup.sh (!)
5. Install appengine from the frankenserver rather than google
6. Remove unused repo-backup from setup.sh; we use phabricator for
   this now
7. Remove some random hg stuff from setup.sh
8. Add a canonical place on toby for the alertlib secret, so many
   clients can use alertlib without needing the secret separately.
9. It's khan, not kahn!  Geesh.

I also fix a longstanding issue with the 'make symlinks' feature of
jenkins, that it didn't delete obsolete symlinks before.

Note that (1) requires changes to the beep-boop and
gae-default-version-notifier repos, which will have to be committed in
parallel with this change.

Add alertlib to the path as well.

I'll need this since the repo isn't really set up to just
clone-and-use (the top-level alertlib dir doesn't have an
__init__.py).

Increase a cronjob timeout, and correctly handle submodules.

I've added submodules to the beep-boop repo among others.  But I was
forgetting to initialize them in setup.sh!  Now I am.  Just to be
safe, I *always* initialize submodules when doing a 'git pull' in
these setup scripts.

Chris also pointed out that bigquery querying can take a little while,
so up the timeout on that.

Fix the install_appengine command.

I was pulling from the wrong directory.

Fix _install_alertlib_secret to make the needed directory as well.

Test Plan:
Fingers crossed.  I haven't run any of these commands yet, but will do
so between the time this change is approved and when I land it.  (That
way I won't run any commands whose ideas don't pass review muster.)

(same)

(same)

Ran it manually on toby.

Ran it by hand on toby.

Reviewers: chris

Reviewed By: chris

Subscribers: alan

Differential Revision: http://phabricator.khanacademy.org/D11438