Get varnish to optionally pull from a sub-domain, if it exists.
Summary: This makes use of the new sub-domain way of setting the host to pull from, in our Varnish server. I left in the old `host=` way of getting the host, until we transition completely to this new way.

Test Plan: Confirmed that the syntax is correct (by running `varnishd -C -f` on the file). Haven't tested this on the live site.

Reviewers: alpert

Reviewed By: alpert

Subscribers: pamela

Differential Revision: http://phabricator.khanacademy.org/D12077