Update the innodb buffer poool size.
Summary:
This is based on a config issue report from phabricator:
   http://phabricator.khanacademy.org/config/issue/mysql.innodb_buffer_pool_size/

```
Your MySQL instance is configured with a very small innodb_buffer_pool_size (128 MB). This may cause poor database performance and lock exhaustion.

There are no hard-and-fast rules to setting an appropriate value, but a reasonable starting point for a standard install is something like 40% of the total memory on the machine. For example, if you have 4GB of RAM on the machine you have installed Phabricator on, you might set this value to 1600M.

You can read more about this option in the MySQL documentation to help you make a decision about how to configure it for your use case. There are no concerns specific to Phabricator which make it different from normal workloads with respect to this setting.

To adjust the setting, add something like this to your my.cnf file (in the [mysqld] section), replacing 1600M with an appropriate value for your host and use case. Then restart mysqld:

innodb_buffer_pool_size=1600M

If you're satisfied with the current setting, you can safely ignore this setup warning.
```

Heck, I want a faster mysql!  I have no idea if this will make a
difference in our case, but it can't hurt.

See, e.g.
    http://www.percona.com/blog/2007/11/03/choosing-innodb_buffer_pool_size/
for more details.

We tend to run a lot of php instances on this machine, and they take
up a lot of memory too, so I won't go with the suggested 1600M, but
1024M to start off with.

Test Plan: Fingers crossed.

Reviewers: alpert

Reviewed By: alpert

Differential Revision: http://phabricator.khanacademy.org/D13697