Send mail when the convert job fails.
It's been failing for the last two months (due to a bad ssl.cert file)
and we hadn't noticed.  Hopefully now we will, or at least will come
closer to.

toby-admin is not the best place to send this kind of mail, but I
don't know of any place better so it will have to do.

Auditors: alpert

Test plan:
Mostly fingers crossed.  I did send myself mail from the
youtube-export machine ('mail csilvers@khanacademy.org') and it got
through, so that's good.