Fix this to have only the content it should.
When I renamed this from nginx.conf, I thought it was just a matter of
renaming, but it's not: the nginx 'sub-configurations' just have the
contents of the 'http' section, not the other config sections like
'event' and the global config.

I just take out those other things, leaving the stuff that used to be
inside http.

Auditors: alpert

Test plan:
Will deploy on caching-proxy and run 'service nginx start' and hope
for the best...