Add error-monitor-db support to the prod-rpc machine.
Summary:
Next time I run setup.sh, if all goes well it will start the
error-monitor-db server.

While in the area I added a host-checking test for nginx, to ensure
that, at setup time, we've set up all the hostnames we'll need to use
the services on a machine successfully.

Note I did not update vagrant_data/preinstall.sh and postinstall.sh,
since I didn't know the best way to do that.  I seemed to be
reproducing a lot of the code already in setup.sh; the two methods may
just not mix very well.  (I already had to do the nginx host-checking
logic in two places, to support vagrant.)  I'll leave it be for now,
but if a vagrant expert wants to improve this I'm all ears.

Test Plan:
Fingers crossed.  Once this is checked in, I'll try running it on the
prod-rpc machine, and then will probably be making changes to get
things working.

Reviewers: tom, johnsullivan

Reviewed By: tom, johnsullivan

Differential Revision: http://phabricator.khanacademy.org/D14532