Don't call 'npm update' anymore.
It definitely shouldn't be called in a utility function.  It should
possibly be called the same place 'npm install' is called, in order to
get new versions of packages if needed.  But that's kinda dangerous --
it could break things that used to work -- so I'm just leaving off the
updating for now.

Auditors: chris

Test plan:
Fingers crossed