Restart cron after changing timezones.
We were seeing problems on toby with cronjobs running at the wrong
time.  My theory is this is because cron was never restarted after the
timezone changed, so it was still basing its decisions on the old
timezone data.  That's just a guess though; we'll see!  Restarting
cron doesn't hurt in any case.

Auditors: alan

Test plan:
Fingers crossed