Avoid a git misfeature that caused continual git repacking.
Summary:
Also try to avoid future trouble by making the git gc-ing
synchronous.  This way, it may take a lot of memory, but won't
interfere with update-translations (say), which also takes a lot of
memory.

Full description of what went wrong and how this fixes it:
   https://docs.google.com/a/khanacademy.org/document/d/1SqJbJBMH3rBKGi-baMkkxcufwnAyEGmHfLEvYGU9HF4/edit#

Test Plan: Fingers crossed!

Reviewers: james, chris

Reviewed By: chris

Differential Revision: https://phabricator.khanacademy.org/D15165