Move logs to ephemeral disk, and make sure nginx logs there.
Summary:
We had a problem today with the caching-proxy machine running out of
space on / due to nginx logs.  This should help ensure that doesn't
happen again.

We now automatically set up a logrotate for these nginx logs as well,
as part of the nginx setup process.  (nginx sets up a logrotate for
logs in the default location, but we need to set up our own since we
put the logs in a different location.)

While in the area, I also automatically set up logs symlinks for
upstart files, into our new logs directory.

Test Plan:
Fingers crossed.  Will try running this stuff on each of these
machines after it's checked in.

Reviewers: alpert, chris

Reviewed By: chris

Subscribers: john, pamela

Differential Revision: https://phabricator.khanacademy.org/D15217