Add a cron job to check on diskspace, and make sure it mails on failure.
Summary:
The mailing part was the more difficult.  I set up /etc/aliases so the
default cron activity would result in mail being sent out, and created
the necessary mailing lists in google groups.

Warn at 90% full, not 95%, and save the warning in the complain-file.

Test Plan:
Fingers crossed.  I'll install this on all the machines and see how it
does. :-)  I may try manually lowering the threshold on one of the
machines to test that mailing works properly.

(same)

Reviewers: alpert, chris

Reviewed By: chris

Subscribers: pamela

Differential Revision: https://phabricator.khanacademy.org/D15222