Install 'git new-workdir' on jenkins.
This is a script we're going to use to share git objects between
different workspaces, saving time and space.  cf.
   http://nuclearsquid.com/writings/git-new-workdir/

I also got rid of a call to a function that doesn't exist.  Dunno
where it came from.

Auditors: chris

Test plan:
Ran the two new lines manually on jenkins, and then ran
   % git new-workdir
   usage: /usr/local/bin/git-new-workdir <repository> <new_workdir> [<branch>]