Add a few more mysql options, as suggested by phabricator.
Summary:
I now also put in comments explaining what they're for.

Once this is committed, I'll need to git pull on phabricator machine
(and maybe copy the file over to /etc -- have to see if it's a symlink
or not).  Then I'll need to run

   mysql> REPAIR TABLE phabricator_search.search_documentfield;

to have the new search options take effect.

Test Plan: Fingers crossed.

Reviewers: chris

Reviewed By: chris

Differential Revision: https://phabricator.khanacademy.org/D15776