Rotate the php5 log file.
Summary:
As far as I can tell, this log file is created by the php5-fpm dpkg,
but they didn't add in a logrotate rule as well, the dastards.  So we
have to do one ourselves.

I ran
   find /var/log -ls | sort -k7n
to verify there were no other big logs being created on phabricator
that weren't being rotated.  (I did the same on toby and jenkins.)
This was the only one.

Test Plan: Fingers crossed.

Reviewers: chris

Reviewed By: chris

Differential Revision: https://phabricator.khanacademy.org/D15829