Add an rc.local to every repo that creates /mnt/logs.
Summary:
We store logs on ephemeral disk, in a directory that we create
(/mnt/logs).  This works fine the first time you set up a machine, but
when it reboots, the ephemeral disk gets wiped (of course), including
/mnt/logs itself!  We set up that directory again, now, on startup.

I also took the opportunity to turn on ptrace support at startup,
since it's really annoying that it's disabled by default in ubuntu.
I used to do this just on toby, now it's done everywhere.

I also fixed a bug in the existing rc.local scripts where they were
being ignored because they did not have the executable bit set.

Test Plan: Fingers crossed.

Reviewers: chris

Reviewed By: chris

Differential Revision: https://phabricator.khanacademy.org/D15830